third_party/blink/web_tests/platform/win/external/wpt/editing/run/subscript-expected.txt

This is a testharness.js-based test.
Found 244 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["subscript",""]] "foo[]bar" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""]] "foo[]bar" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["subscript",""]] "<p>[foo</p> <p>bar]</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><sub>foo</sub></p> <p><sub>bar</sub></p>" but got "<p><span style=\\"vertical-align:sub\\">foo</span></p> <p><span style=\\"vertical-align:sub\\">bar</span></p>"
[FAIL] [["stylewithcss","true"],["subscript",""]] "<p>[foo</p> <p>bar]</p>" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["subscript",""]] "<p>[foo</p> <p>bar]</p>" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["subscript",""]] "<p>[foo</p> <p>bar]</p>" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["subscript",""]] "<p>[foo</p> <p>bar]</p>" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["subscript",""]] "<p>[foo</p> <p>bar]</p>" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["subscript",""]] "<span>[foo</span> <span>bar]</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<sub><span>foo</span> <span>bar</span></sub>" but got "<span style=\\"vertical-align:sub\\"><span>foo</span> <span>bar</span></span>"
[FAIL] [["stylewithcss","true"],["subscript",""]] "<span>[foo</span> <span>bar]</span>" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["subscript",""]] "<span>[foo</span> <span>bar]</span>" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["subscript",""]] "<span>[foo</span> <span>bar]</span>" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["subscript",""]] "<span>[foo</span> <span>bar]</span>" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["subscript",""]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><sub>foo</sub></p><p> <sub><span>bar</span></sub> </p><p><sub>baz</sub></p>" but got "<p><span style=\\"vertical-align:sub\\">foo</span></p><p> <span style=\\"vertical-align:sub\\"><span>bar</span> </span></p><p><span style=\\"vertical-align:sub\\">baz</span></p>"
[FAIL] [["stylewithcss","true"],["subscript",""]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["subscript",""]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["subscript",""]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><sub>foo</sub></p><p> <sub><span>bar</span></sub> </p><p><sub>baz</sub></p>" but got "<p><sub>foo</sub></p><p> <sub><span>bar</span> </sub></p><p><sub>baz</sub></p>"
[FAIL] [["stylewithcss","false"],["subscript",""]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["subscript",""]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["subscript",""]] "<p>[foo<p><br><p>bar]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><sub>foo</sub></p><p><sub><br></sub></p><p><sub>bar</sub></p>" but got "<p><span style=\\"vertical-align:sub\\">foo</span></p><p><span style=\\"vertical-align:sub\\"><br></span></p><p><span style=\\"vertical-align:sub\\">bar</span></p>"
[FAIL] [["stylewithcss","true"],["subscript",""]] "<p>[foo<p><br><p>bar]" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["subscript",""]] "<p>[foo<p><br><p>bar]" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["subscript",""]] "<p>[foo<p><br><p>bar]" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["subscript",""]] "<p>[foo<p><br><p>bar]" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["subscript",""]] "<b>foo[]bar</b>" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""]] "<b>foo[]bar</b>" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["subscript",""]] "<i>foo[]bar</i>" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""]] "<i>foo[]bar</i>" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["subscript",""]] "<span>foo</span>{}<span>bar</span>" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""]] "<span>foo</span>{}<span>bar</span>" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["subscript",""]] "<span>foo[</span><span>]bar</span>" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""]] "<span>foo[</span><span>]bar</span>" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo[bar]baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sub>bar</sub>baz" but got "foo<span style=\\"vertical-align:sub\\">bar</span>baz"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo[bar]baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo[bar]baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo[bar]baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo[bar]baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo[bar<b>baz]qoz</b>quz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sub>bar</sub><b><sub>baz</sub>qoz</b>quz" but got "foo<span style=\\"vertical-align:sub\\">bar</span><b><span style=\\"vertical-align:sub\\">baz</span>qoz</b>quz"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo[bar<b>baz]qoz</b>quz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo[bar<b>baz]qoz</b>quz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo[bar<b>baz]qoz</b>quz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo[bar<b>baz]qoz</b>quz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo[bar<i>baz]qoz</i>quz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sub>bar</sub><i><sub>baz</sub>qoz</i>quz" but got "foo<span style=\\"vertical-align:sub\\">bar</span><i><span style=\\"vertical-align:sub\\">baz</span>qoz</i>quz"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo[bar<i>baz]qoz</i>quz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo[bar<i>baz]qoz</i>quz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo[bar<i>baz]qoz</i>quz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo[bar<i>baz]qoz</i>quz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["subscript",""]] "{<p><p> <p>foo</p>}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p></p><p> </p><p><sub>foo</sub></p>" but got "<p></p><p> </p><p><span style=\\"vertical-align:sub\\">foo</span></p>"
[FAIL] [["stylewithcss","true"],["subscript",""]] "{<p><p> <p>foo</p>}" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["subscript",""]] "{<p><p> <p>foo</p>}" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["subscript",""]] "{<p><p> <p>foo</p>}" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["subscript",""]] "{<p><p> <p>foo</p>}" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["subscript",""]] "<table><tbody><tr><td>foo<td>b[a]r<td>baz</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td>foo</td><td>b<sub>a</sub>r</td><td>baz</td></tr></tbody></table>" but got "<table><tbody><tr><td>foo</td><td>b<span style=\\"vertical-align:sub\\">a</span>r</td><td>baz</td></tr></tbody></table>"
[FAIL] [["stylewithcss","true"],["subscript",""]] "<table><tbody><tr><td>foo<td>b[a]r<td>baz</table>" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["subscript",""]] "<table><tbody><tr><td>foo<td>b[a]r<td>baz</table>" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["subscript",""]] "<table><tbody><tr><td>foo<td>b[a]r<td>baz</table>" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["subscript",""]] "<table><tbody><tr><td>foo<td>b[a]r<td>baz</table>" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["subscript",""]] "<table><tbody><tr data-start=1 data-end=2><td>foo<td>bar<td>baz</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td>foo</td><td><sub>bar</sub></td><td>baz</td></tr></tbody></table>" but got "<table><tbody><tr><td>foo</td><td><span style=\\"vertical-align:sub\\">bar</span></td><td>baz</td></tr></tbody></table>"
[FAIL] [["stylewithcss","true"],["subscript",""]] "<table><tbody><tr data-start=1 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["subscript",""]] "<table><tbody><tr data-start=1 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["subscript",""]] "<table><tbody><tr data-start=1 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["subscript",""]] "<table><tbody><tr data-start=1 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["subscript",""]] "<table><tbody><tr data-start=0 data-end=2><td>foo<td>bar<td>baz</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td><sub>foo</sub></td><td><sub>bar</sub></td><td>baz</td></tr></tbody></table>" but got "<table><tbody><tr><td><span style=\\"vertical-align:sub\\">foo</span></td><td><span style=\\"vertical-align:sub\\">bar</span></td><td>baz</td></tr></tbody></table>"
[FAIL] [["stylewithcss","true"],["subscript",""]] "<table><tbody><tr data-start=0 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["subscript",""]] "<table><tbody><tr data-start=0 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["subscript",""]] "<table><tbody><tr data-start=0 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["subscript",""]] "<table><tbody><tr data-start=0 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["subscript",""]] "<table><tbody data-start=0 data-end=1><tr><td>foo<td>bar<td>baz</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td><sub>foo</sub></td><td><sub>bar</sub></td><td><sub>baz</sub></td></tr></tbody></table>" but got "<table><tbody><tr><td><span style=\\"vertical-align:sub\\">foo</span></td><td><span style=\\"vertical-align:sub\\">bar</span></td><td><span style=\\"vertical-align:sub\\">baz</span></td></tr></tbody></table>"
[FAIL] [["stylewithcss","true"],["subscript",""]] "<table><tbody data-start=0 data-end=1><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["subscript",""]] "<table><tbody data-start=0 data-end=1><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["subscript",""]] "<table><tbody data-start=0 data-end=1><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["subscript",""]] "<table><tbody data-start=0 data-end=1><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["subscript",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td><sub>foo</sub></td><td><sub>bar</sub></td><td><sub>baz</sub></td></tr></tbody></table>" but got "<table><tbody><tr><td><span style=\\"vertical-align:sub\\">foo</span></td><td><span style=\\"vertical-align:sub\\">bar</span></td><td><span style=\\"vertical-align:sub\\">baz</span></td></tr></tbody></table>"
[FAIL] [["stylewithcss","true"],["subscript",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["subscript",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["subscript",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["subscript",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["subscript",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["subscript",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["subscript",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td><sub>foo</sub></td><td><sub>bar</sub></td><td><sub>baz</sub></td></tr></tbody></table>" but got "<table><tbody><tr><td><span style=\\"vertical-align:sub\\">foo</span></td><td><span style=\\"vertical-align:sub\\">bar</span></td><td><span style=\\"vertical-align:sub\\">baz</span></td></tr></tbody></table>"
[FAIL] [["stylewithcss","true"],["subscript",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["subscript",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["subscript",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["subscript",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["subscript",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["subscript",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sub>b[a]r</sub>baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sub>b[a]r</sub>baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sub>b[a]r</sub>baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sub>b[a]r</sub>baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sup>[bar]</sup>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sub>bar</sub>baz" but got "foo<span style=\\"vertical-align:sub\\">bar</span>baz"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sup>[bar]</sup>baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sup>[bar]</sup>baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sup>[bar]</sup>baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sup>[bar]</sup>baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sup>b[a]r</sup>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sup>b</sup><sub>a</sub><sup>r</sup>baz" but got "foo<sup>b</sup><span style=\\"vertical-align:sub\\">a</span><sup>r</sup>baz"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sup>b[a]r</sup>baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sup>b[a]r</sup>baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sup>b[a]r</sup>baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sup>b[a]r</sup>baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<span style=vertical-align:sub>[bar]</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sub>bar</sub>baz" but got "foobarbaz"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<span style=vertical-align:sub>[bar]</span>baz" queryCommandState("subscript") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<span style=vertical-align:sub>[bar]</span>baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<span style=vertical-align:sub>[bar]</span>baz" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<span style=vertical-align:sub>[bar]</span>baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<span style=vertical-align:sub>[bar]</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sub>bar</sub>baz" but got "foobarbaz"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<span style=vertical-align:sub>[bar]</span>baz" queryCommandState("subscript") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<span style=vertical-align:sub>[bar]</span>baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<span style=vertical-align:sub>[bar]</span>baz" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<span style=vertical-align:sub>[bar]</span>baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<span style=vertical-align:super>[bar]</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sub>bar</sub>baz" but got "foo<span style=\\"vertical-align:sub\\">bar</span>baz"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<span style=vertical-align:super>[bar]</span>baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<span style=vertical-align:super>[bar]</span>baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<span style=vertical-align:super>[bar]</span>baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<span style=vertical-align:super>[bar]</span>baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sub><sub>[bar]</sub></sub>baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sub><sub>[bar]</sub></sub>baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sub><sub>[bar]</sub></sub>baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sub><sub>[bar]</sub></sub>baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""]] "foo<sub><sub>b[a]r</sub></sub>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sub>b</sub>a<sub>r</sub>baz" but got "foo<sub style=\\"vertical-align:sub\\">b</sub>a<sub style=\\"vertical-align:sub\\">r</sub>baz"
[FAIL] [["subscript",""]] "foo<sub><sub>b[a]r</sub></sub>baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["subscript",""]] "foo<sub><sub>b[a]r</sub></sub>baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""]] "foo<sub>b<sub>[a]</sub>r</sub>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sub>b</sub>a<sub>r</sub>baz" but got "foo<sub style=\\"vertical-align:sub\\">b</sub>a<sub>r</sub>baz"
[FAIL] [["subscript",""]] "foo<sub>b<sub>[a]</sub>r</sub>baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["subscript",""]] "foo<sub>b<sub>[a]</sub>r</sub>baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sup><sup>[bar]</sup></sup>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sub>bar</sub>baz" but got "foo<span style=\\"vertical-align:sub\\">bar</span>baz"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sup><sup>[bar]</sup></sup>baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sup><sup>[bar]</sup></sup>baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sup><sup>[bar]</sup></sup>baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sup><sup>[bar]</sup></sup>baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sup><sup>b[a]r</sup></sup>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sup>b</sup><sub>a</sub><sup>r</sup>baz" but got "foo<sup style=\\"vertical-align:super\\">b</sup><span style=\\"vertical-align:sub\\">a</span><sup style=\\"vertical-align:super\\">r</sup>baz"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sup><sup>b[a]r</sup></sup>baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sup><sup>b[a]r</sup></sup>baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sup><sup>b[a]r</sup></sup>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sup>b</sup><sub>a</sub><sup>r</sup>baz" but got "foo<sup style=\\"vertical-align:super\\">b</sup><sub>a</sub><sup style=\\"vertical-align:super\\">r</sup>baz"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sup><sup>b[a]r</sup></sup>baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sup><sup>b[a]r</sup></sup>baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sup>b<sup>[a]</sup>r</sup>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sup>b</sup><sub>a</sub><sup>r</sup>baz" but got "foo<span style=\\"vertical-align:super\\">b</span><span style=\\"vertical-align:sub\\">a</span><span style=\\"vertical-align:super\\">r</span>baz"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sup>b<sup>[a]</sup>r</sup>baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sup>b<sup>[a]</sup>r</sup>baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sup>b<sup>[a]</sup>r</sup>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sup>b</sup><sub>a</sub><sup>r</sup>baz" but got "foo<sup style=\\"vertical-align:super\\">b</sup><sub>a</sub><sup>r</sup>baz"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sup>b<sup>[a]</sup>r</sup>baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sup>b<sup>[a]</sup>r</sup>baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sub><sup>[bar]</sup></sub>baz" queryCommandIndeterm("subscript") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sub><sup>[bar]</sup></sub>baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sub><sup>[bar]</sup></sub>baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sub><sup>[bar]</sup></sub>baz" queryCommandIndeterm("subscript") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sub><sup>[bar]</sup></sub>baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sub><sup>[bar]</sup></sub>baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sub><sup>b[a]r</sup></sub>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sup>b</sup><sub>a</sub><sup>r</sup>baz" but got "foo<sub><sup>b</sup>a<sup>r</sup></sub>baz"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sub><sup>b[a]r</sup></sub>baz" queryCommandIndeterm("subscript") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sub><sup>b[a]r</sup></sub>baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sub><sup>b[a]r</sup></sub>baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sub><sup>b[a]r</sup></sub>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sup>b</sup><sub>a</sub><sup>r</sup>baz" but got "foo<sub><sup>b</sup>a<sup>r</sup></sub>baz"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sub><sup>b[a]r</sup></sub>baz" queryCommandIndeterm("subscript") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sub><sup>b[a]r</sup></sub>baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sub><sup>b[a]r</sup></sub>baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sub>b<sup>[a]</sup>r</sub>baz" queryCommandIndeterm("subscript") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sub>b<sup>[a]</sup>r</sub>baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sub>b<sup>[a]</sup>r</sub>baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sub>b<sup>[a]</sup>r</sub>baz" queryCommandIndeterm("subscript") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sub>b<sup>[a]</sup>r</sub>baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sub>b<sup>[a]</sup>r</sub>baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sup><sub>[bar]</sub></sup>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sub>bar</sub>baz" but got "foobarbaz"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sup><sub>[bar]</sub></sup>baz" queryCommandIndeterm("subscript") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sup><sub>[bar]</sub></sup>baz" queryCommandState("subscript") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sup><sub>[bar]</sub></sup>baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sup><sub>[bar]</sub></sup>baz" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sup><sub>[bar]</sub></sup>baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sup><sub>[bar]</sub></sup>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sub>bar</sub>baz" but got "foobarbaz"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sup><sub>[bar]</sub></sup>baz" queryCommandIndeterm("subscript") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sup><sub>[bar]</sub></sup>baz" queryCommandState("subscript") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sup><sub>[bar]</sub></sup>baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sup><sub>[bar]</sub></sup>baz" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sup><sub>[bar]</sub></sup>baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sup><sub>b[a]r</sub></sup>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sub>bar</sub>baz" but got "foo<sub style=\\"vertical-align:super\\">b</sub>a<sub style=\\"vertical-align:super\\">r</sub>baz"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sup><sub>b[a]r</sub></sup>baz" queryCommandIndeterm("subscript") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sup><sub>b[a]r</sub></sup>baz" queryCommandState("subscript") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sup><sub>b[a]r</sub></sup>baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sup><sub>b[a]r</sub></sup>baz" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sup><sub>b[a]r</sub></sup>baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sup><sub>b[a]r</sub></sup>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sub>bar</sub>baz" but got "foo<sub style=\\"vertical-align:super\\">b</sub>a<sub style=\\"vertical-align:super\\">r</sub>baz"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sup><sub>b[a]r</sub></sup>baz" queryCommandIndeterm("subscript") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sup><sub>b[a]r</sub></sup>baz" queryCommandState("subscript") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sup><sub>b[a]r</sub></sup>baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sup><sub>b[a]r</sub></sup>baz" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sup><sub>b[a]r</sub></sup>baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sup>b<sub>[a]</sub>r</sup>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sup>b</sup><sub>a</sub><sup>r</sup>baz" but got "foo<span style=\\"vertical-align:super\\">b</span>a<span style=\\"vertical-align:super\\">r</span>baz"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sup>b<sub>[a]</sub>r</sup>baz" queryCommandIndeterm("subscript") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sup>b<sub>[a]</sub>r</sup>baz" queryCommandState("subscript") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sup>b<sub>[a]</sub>r</sup>baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sup>b<sub>[a]</sub>r</sup>baz" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sup>b<sub>[a]</sub>r</sup>baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sup>b<sub>[a]</sub>r</sup>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sup>b</sup><sub>a</sub><sup>r</sup>baz" but got "foo<sup>b</sup>a<sup>r</sup>baz"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sup>b<sub>[a]</sub>r</sup>baz" queryCommandIndeterm("subscript") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sup>b<sub>[a]</sub>r</sup>baz" queryCommandState("subscript") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sup>b<sub>[a]</sub>r</sup>baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sup>b<sub>[a]</sub>r</sup>baz" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sup>b<sub>[a]</sub>r</sup>baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""]] "fo[o<sub>b]ar</sub>baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""]] "fo[o<sub>b]ar</sub>baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["subscript",""]] "foo<sub>ba[r</sub>b]az" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""]] "foo<sub>ba[r</sub>b]az" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["subscript",""]] "fo[o<sub>bar</sub>b]az" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""]] "fo[o<sub>bar</sub>b]az" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["subscript",""]] "foo[<sub>b]ar</sub>baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["subscript",""]] "foo[<sub>b]ar</sub>baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""]] "foo<sub>ba[r</sub>]baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["subscript",""]] "foo<sub>ba[r</sub>]baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo[<sub>bar</sub>]baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo[<sub>bar</sub>]baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo[<sub>bar</sub>]baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo[<sub>bar</sub>]baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sub>[bar]</sub>baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<sub>[bar]</sub>baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sub>[bar]</sub>baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<sub>[bar]</sub>baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo{<sub>bar</sub>}baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo{<sub>bar</sub>}baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo{<sub>bar</sub>}baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo{<sub>bar</sub>}baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""]] "<sub>fo[o</sub><sup>b]ar</sup>" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""]] "<sub>fo[o</sub><sup>b]ar</sup>" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["subscript",""]] "<sub>fo[o</sub><span style=vertical-align:sub>b]ar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<sub>foob</sub><span style=\\"vertical-align:sub\\">ar</span>" but got "<sub>fo</sub>ob<span style=\\"vertical-align:sub\\">ar</span>"
[FAIL] [["stylewithcss","true"],["subscript",""]] "<sub>fo[o</sub><span style=vertical-align:sub>b]ar</span>" queryCommandIndeterm("subscript") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["subscript",""]] "<sub>fo[o</sub><span style=vertical-align:sub>b]ar</span>" queryCommandState("subscript") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["subscript",""]] "<sub>fo[o</sub><span style=vertical-align:sub>b]ar</span>" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["subscript",""]] "<sub>fo[o</sub><span style=vertical-align:sub>b]ar</span>" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["subscript",""]] "<sub>fo[o</sub><span style=vertical-align:sub>b]ar</span>" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["subscript",""]] "<sub>fo[o</sub><span style=vertical-align:sub>b]ar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<sub>foob</sub><span style=\\"vertical-align:sub\\">ar</span>" but got "<sub>fo</sub>ob<span style=\\"vertical-align:sub\\">ar</span>"
[FAIL] [["stylewithcss","false"],["subscript",""]] "<sub>fo[o</sub><span style=vertical-align:sub>b]ar</span>" queryCommandIndeterm("subscript") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["subscript",""]] "<sub>fo[o</sub><span style=vertical-align:sub>b]ar</span>" queryCommandState("subscript") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","false"],["subscript",""]] "<sub>fo[o</sub><span style=vertical-align:sub>b]ar</span>" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["subscript",""]] "<sub>fo[o</sub><span style=vertical-align:sub>b]ar</span>" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["subscript",""]] "<sub>fo[o</sub><span style=vertical-align:sub>b]ar</span>" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["subscript",""]] "<sub>fo[o</sub><span style=vertical-align:top>b]ar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<sub>foob</sub><span style=\\"vertical-align:top\\">ar</span>" but got "<sub>foo</sub><span style=\\"vertical-align:sub\\">b</span><span style=\\"vertical-align:top\\">ar</span>"
[FAIL] [["stylewithcss","true"],["subscript",""]] "<sub>fo[o</sub><span style=vertical-align:top>b]ar</span>" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["subscript",""]] "<sub>fo[o</sub><span style=vertical-align:top>b]ar</span>" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["subscript",""]] "<sub>fo[o</sub><span style=vertical-align:top>b]ar</span>" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["subscript",""]] "<sub>fo[o</sub><span style=vertical-align:top>b]ar</span>" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<span style=vertical-align:top>[bar]</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sub>bar</sub>baz" but got "foo<span style=\\"vertical-align:sub\\">bar</span>baz"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<span style=vertical-align:top>[bar]</span>baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["subscript",""]] "foo<span style=vertical-align:top>[bar]</span>baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<span style=vertical-align:top>[bar]</span>baz" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["subscript",""]] "foo<span style=vertical-align:top>[bar]</span>baz" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
Harness: the test ran to completion.
